python_onedrive_bridge/requirements.txt
requests==2.31.0
msal==1.24.1
openpyxl==3.1.2
pandas==2.1.4
schedule==1.2.0
python-dotenv==1.0.0


